+++ src/interface/ui.py
@@ -1,9 +1,10 @@
-from rich import print
+from rich.console import Console
+console = Console()
 
 def get_banner() -> str:
-    " Return the content from banner.txt file as application banner "
+    " Return Juumla's banner "
 
-    path: str = "src/interface/banner.txt"
-    with open(path) as file:
-        lines: str = file.read()
-        print(f"[bold yellow]{lines}[/]")
+    console.print("""[bold yellow]
+jUuMlA - 0.1.6
+most overrated joomla scanner
+[/]""")