+++ src/juumla/modules/vulns.py
@@ -1,12 +1,14 @@
 from json import load
-from rich import print
+from rich.console import Console
+console = Console()
 
 from src.juumla.modules.files import files_manager
+
 
 def vuln_manager(url, version) -> None:
     " Search for vulnerabilities on current target Joomla version "
 
-    print("\n[cyan]> Running Joomla vulnerabilities scanner... [2/3] [/]")
+    console.print("\n[yellow][!][/] Running Joomla vulnerabilities scanner! [cyan](2/3)[/]", highlight=False)
 
     filename: str = "src/juumla/data/vulnerabilites.json"
     
@@ -19,10 +21,9 @@ def vuln_manager(url, version) -> None:
             max_version = vuln_version[1]
 
             if version >= min_version and version <= max_version:
-                print(f"[green]> {vuln_title} [/]")
+                console.print(f"[green][+][/] {vuln_title}", highlight=False)
             else:
                 pass
 
-
-    print("[yellow]> Vulnerabilities scanner finished [2/3] [/]")
+    console.print("[yellow][!][/] Vulnerabilities scanner finished! [cyan](2/3)[/]", highlight=False)
     files_manager(url)